Refactor password recovery
