+++ src/app/auth/password/index.tsx
@@ -1,36 +1,40 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { View, Text, ScrollView } from 'react-native'
 import { useForm } from 'react-hook-form'
-import { router } from 'expo-router'
+import { router, useNavigation } from 'expo-router'
 import { zodResolver } from '@hookform/resolvers/zod'
 import Icons from '@/src/components/Icons'
 import Screen from '@/src/components/Screen'
-import { ResetPasswordSchema } from '@/src/schemas/UserSchema'
+import { EmailSchema } from '@/src/schemas/UserSchema'
 import Input from '@/src/components/Input'
 import Button from '@/src/components/Button'
 
-export default function ResetPassword() {
-  const { LockIcon } = Icons
+export default function RecoveryPassword() {
+  const navigation = useNavigation();
+  useEffect(() => {
+    navigation.setOptions({ headerShown: false });
+  }, [navigation]);
+  const { EmailIcon } = Icons
   const { control, handleSubmit, reset } = useForm({
-    resolver: zodResolver(ResetPasswordSchema)
+    resolver: zodResolver(EmailSchema)
   })
 
   return (
     <ScrollView>
       <Screen>
         <View className="flex justify-center items-center">
-          <Text className="text-helper font-pbold text-[20px] mb-5">
-            ¡Hora de cambiar la contraseña!
-          </Text>
-          <LockIcon />
+          <Text className="text-helper font-pbold text-[20px]">Introduce tu dirección de</Text>
+          <Text className="text-helper font-pbold text-[20px] mb-5">correo electrónico</Text>
+          <EmailIcon />
           <Text className="text-lightc font-pbold text-[16px] mt-5">
-            Introduce la nueva contraseña
+            Para recuperar tu contraseña
           </Text>
-          <Input variant="password" control={control} name="newPassword">
-            Contraseña
-          </Input>
-          <Input variant="password" control={control} name="confirmNewPassword">
-            Repite Contraseña
+          <Text className="text-lightc font-pbold text-[16px]">
+            necesitarás el correo electrónico
+          </Text>
+          <Text className="text-lightc font-pbold text-[16px]">vinculado a tu cuenta</Text>
+          <Input variant="email" control={control} name="email">
+            Correo vinculado
           </Input>
         </View>
         <View className="flex flex-row justify-around mt-20">
@@ -40,7 +44,7 @@ export default function ResetPassword() {
             variant="secondary"
             onPress={() => {
               reset()
-              router.replace('/auth/login')
+              router.back()
             }}
           >
             Cancelar
@@ -52,11 +56,11 @@ export default function ResetPassword() {
             onPress={handleSubmit((data) => {
               if (data) {
                 reset()
-                router.push('/(tabs)/Home')
+                router.push('/auth/password/otp')
               }
             })}
           >
-            Actualizar
+            Siguiente
           </Button>
         </View>
       </Screen>
